datatype Long in database changed on int add NewNoteActivity add button and function to add new note to database

+++ app/src/main/java/com/example/dreamsdiary/dao/Label_NotesDao.java
@@ -11,10 +11,10 @@ import java.util.List;
 public interface Label_NotesDao {
 
     @Query("SELECT * FROM Labels WHERE id = (SELECT labelId FROM Label_Notes WHERE noteId = :id)")
-    Labels getAllLabelsForNote(Long id);
+    Labels getAllLabelsForNote(int id);
 
     @Query("SELECT * FROM Notes WHERE id = (SELECT noteId FROM Label_Notes WHERE labelId = :id)")
-    Notes getAllNotesForLabel(Long id);
+    Notes getAllNotesForLabel(int id);
 
     @Insert
     void insert(Label_Notes label_notes);
